Rename JaggedArrayPrep to JaggedArray
Fix a long-standing weird issue with the array implementation stuff
